refactor(chat): migrate selectedRoom middleware to saga
