--- src/middlewares/chatx/index.js
@@ -1,9 +1,0 @@
-import { pipeMiddlewares } from 'redux-fun';
-
-import createSelectedRoomMiddleware from './selectedRoom';
-
-const createChatxMiddleware = api => pipeMiddlewares(
-  createSelectedRoomMiddleware(api),
-);
-
-export default createChatxMiddleware;
