+++ src/store/commonMiddlewares.js
@@ -4,12 +4,8 @@ import { hashHistory } from "react-router";
 
 import sagaMiddleware from '../sagas/middleware';
 
-import createChatxMiddleware from './chatx';
-import firebaseApi from '../api/firebase';
-
 export default [
   thunk,
-  createChatxMiddleware(firebaseApi),
   routerMiddleware(hashHistory),
   sagaMiddleware,
 ];
