+++ src/store/configureStore.dev.js
@@ -2,8 +2,8 @@ import { createStore, applyMiddleware, compose } from 'redux';
 import reduxImmutableStateInvariant from 'redux-immutable-state-invariant';
 import logger from 'redux-logger';
 
+import commonMiddewares from './commonMiddlewares';
 import rootReducer from '../reducers';
-import middlewares from '../middlewares';
 import runSagas from '../sagas';
 
 
@@ -13,7 +13,7 @@ export default function configureStore(initialState) {
     initialState,
     compose(
       applyMiddleware(
-        ...middlewares,
+        ...commonMiddewares,
         reduxImmutableStateInvariant(),
         logger,
       ),
